Update UI script namespaces
